move storage to pkg
