+++ server/cluster/cluster_worker_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/pd/pkg/core"
 	"github.com/tikv/pd/pkg/mock/mockid"
-	"github.com/tikv/pd/server/storage"
+	"github.com/tikv/pd/pkg/storage"
 )
 
 func TestReportSplit(t *testing.T) {
