+++ server/cluster/unsafe_recovery_controller_test.go
@@ -27,8 +27,8 @@ import (
 	"github.com/tikv/pd/pkg/codec"
 	"github.com/tikv/pd/pkg/core"
 	"github.com/tikv/pd/pkg/mock/mockid"
+	"github.com/tikv/pd/pkg/storage"
 	"github.com/tikv/pd/server/schedule/hbstream"
-	"github.com/tikv/pd/server/storage"
 )
 
 func newStoreHeartbeat(storeID uint64, report *pdpb.StoreReport) *pdpb.StoreHeartbeatRequest {
