+++ server/schedulers/balance_witness_test.go
@@ -20,11 +20,11 @@ import (
 
 	"github.com/stretchr/testify/suite"
 	"github.com/tikv/pd/pkg/mock/mockcluster"
+	"github.com/tikv/pd/pkg/storage"
 	"github.com/tikv/pd/server/config"
 	"github.com/tikv/pd/server/schedule"
 	"github.com/tikv/pd/server/schedule/operator"
 	"github.com/tikv/pd/server/schedule/placement"
-	"github.com/tikv/pd/server/storage"
 )
 
 func TestBalanceWitnessSchedulerTestSuite(t *testing.T) {
