+++ tests/server/region_syncer/region_syncer_test.go
@@ -48,7 +48,7 @@ func TestRegionSyncer(t *testing.T) {
 	re := require.New(t)
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	re.NoError(failpoint.Enable("github.com/tikv/pd/server/storage/regionStorageFastFlush", `return(true)`))
+	re.NoError(failpoint.Enable("github.com/tikv/pd/pkg/storage/regionStorageFastFlush", `return(true)`))
 	re.NoError(failpoint.Enable("github.com/tikv/pd/server/syncer/noFastExitSync", `return(true)`))
 
 	cluster, err := tests.NewTestCluster(ctx, 3, func(conf *config.Config, serverName string) { conf.PDServerCfg.UseRegionStorage = true })
@@ -152,7 +152,7 @@ func TestRegionSyncer(t *testing.T) {
 		re.Equal(region.GetBuckets(), r.GetBuckets())
 	}
 	re.NoError(failpoint.Disable("github.com/tikv/pd/server/syncer/noFastExitSync"))
-	re.NoError(failpoint.Disable("github.com/tikv/pd/server/storage/regionStorageFastFlush"))
+	re.NoError(failpoint.Disable("github.com/tikv/pd/pkg/storage/regionStorageFastFlush"))
 }
 
 func TestFullSyncWithAddMember(t *testing.T) {
